chore(*): minor tidy up and add snapshot tests for ODRL templates
