+++ frontend/__tests__/fixtures/RA61D4c7-odrl-policy-template_output.trig
@@ -1,0 +1,55 @@
+@prefix this: <https://w3id.org/sciencelive/np/RAzkXH-E0fxztdJbff26ZJC02dkYq_Tffs9KqN8a-ql2s>.
+@prefix sub: <https://w3id.org/sciencelive/np/RAzkXH-E0fxztdJbff26ZJC02dkYq_Tffs9KqN8a-ql2s/>.
+@prefix np: <http://www.nanopub.org/nschema#>.
+@prefix rdf: <http://www.w3.org/1999/02/22-rdf-syntax-ns#>.
+@prefix prov: <http://www.w3.org/ns/prov#>.
+@prefix npx: <http://purl.org/nanopub/x/>.
+@prefix dc: <http://purl.org/dc/terms/>.
+@prefix xsd: <http://www.w3.org/2001/XMLSchema#>.
+
+sub:Head {
+this: np:hasAssertion sub:assertion;
+    np:hasProvenance sub:provenance;
+    np:hasPublicationInfo sub:pubinfo;
+    a np:Nanopublication
+}
+sub:assertion {
+<constNode> <http://www.w3.org/ns/odrl/2/leftOperand> <http://www.w3.org/ns/odrl/2/purpose>;
+    <http://www.w3.org/ns/odrl/2/operator> <http://www.w3.org/ns/odrl/2/eq>;
+    <http://www.w3.org/ns/odrl/2/rightOperand> <https://w3id.org/dpv#AcademicResearch>.
+<constNode__1> <http://www.w3.org/ns/odrl/2/leftOperand> <http://www.w3.org/ns/odrl/2/purpose>;
+    <http://www.w3.org/ns/odrl/2/operator> <http://www.w3.org/ns/odrl/2/eq>;
+    <http://www.w3.org/ns/odrl/2/rightOperand> <https://w3id.org/dpv#AcademicResearch>.
+<dutyNode> <http://www.w3.org/ns/odrl/2/action> <http://www.w3.org/ns/odrl/2/attribute>;
+    <http://www.w3.org/ns/odrl/2/attributedParty> <https://fair2adapt-eosc.eu>.
+<https://fair2adapt.eu/policy/public-demo-biodiversity> a <http://www.w3.org/ns/odrl/2/Offer>;
+    <http://www.w3.org/ns/odrl/2/duty> <dutyNode>;
+    <http://www.w3.org/ns/odrl/2/permission> <permNode>, <permNode__1>;
+    <http://www.w3.org/ns/odrl/2/prohibition> <prohibNode>, <prohibNode__1>;
+    <http://www.w3.org/ns/odrl/2/target> <https://fair2adapt.eu/data/public-demo-biodiversity>.
+<permNode> <http://www.w3.org/ns/odrl/2/action> <http://www.w3.org/ns/odrl/2/use>;
+    <http://www.w3.org/ns/odrl/2/constraint> <constNode>.
+<permNode__1> <http://www.w3.org/ns/odrl/2/action> <http://www.w3.org/ns/odrl/2/reproduce>;
+    <http://www.w3.org/ns/odrl/2/constraint> <constNode__1>.
+<prohibNode> <http://www.w3.org/ns/odrl/2/action> <http://www.w3.org/ns/odrl/2/commercialize>.
+<prohibNode__1> <http://www.w3.org/ns/odrl/2/action> <http://www.w3.org/ns/odrl/2/sell>
+}
+sub:provenance {
+sub:assertion prov:wasAttributedTo <https://orcid.org/0000-0000-0000-0000>
+}
+sub:pubinfo {
+this: dc:created "2026-04-11T12:00:00.000Z"^^xsd:dateTime;
+    dc:creator <https://orcid.org/0000-0000-0000-0000>;
+    dc:license <https://creativecommons.org/licenses/by/4.0/>;
+    npx:hasNanopubType <http://www.w3.org/ns/odrl/2/Policy>;
+    npx:introduces <https://fair2adapt.eu/policy/public-demo-biodiversity>;
+    npx:wasCreatedAt <https://platform.sciencelive4all.org>;
+    <http://www.w3.org/2000/01/rdf-schema#label> "ODRL policy: public-demo-biodiversity";
+    <https://w3id.org/np/o/ntemplate/wasCreatedFromTemplate> <https://w3id.org/np/RA61D4c7dB5t0B1mLhc78bN2vagqYTXQiJDKY0yImRULI>.
+<https://orcid.org/0000-0000-0000-0000> <http://xmlns.com/foaf/0.1/name> "Test User".
+sub:sig npx:signedBy <https://orcid.org/0000-0000-0000-0000>;
+    npx:hasAlgorithm "RSA";
+    npx:hasPublicKey "MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAzUBHKF13IXaU5u9U4nYcq9QTCafI2HlYO/6C25hOVjidIYpUh/25or8sAfh6Ezq+v97aYp8NAcLNjNpv+yy5w2mnmAeUQtg6A9gSgWuPfcjh2t6xjBjof74sghjiW4xTyDZTIbCM7K/sL5LR8NCXJX1udPJR5rnY80df+gdzj3W4+l8NvdlRyEbdlncgpjxHPP98787snEN8bFEZzMUaAHRaiH+Yj99KZ/bqj/hMda1GmeJW61PALjT8rN34iDw1U2ILIeGpWPyHGL3eTgr59rsOC6F3fyWW/RCbktSmCvb/w0YqwnZb0jYeIcuVBO/5zYJlw7d8WIw38zmDf/DfYQIDAQAB";
+    npx:hasSignatureTarget this:;
+    npx:hasSignature "IveVdn/toqDPonrJ1rvE4EqSBSWUSNAEI2yHDBH+6D2OoOmVKclknntMrXbJY8VUSORust/Qu7cD/Nk4nDNfDsnwfB21QClFIYmBe3p4RIaAN/mpozKJm+xpelh46Ks6HDG8UzjJEtXNbbC8I9KS3gWloxlMP8pF90C+H8maAPG3sRxnE5F8IAs49XvUMb3YdkPERD/CsSQehH1u0umeCWFI84WGl6vQ2XfwXAsjis5amOAHnRpU+xJtKYpE2WzjhbAXFVwQfg/kt8xT0rvO9GzKCJtCJpWaNLx0NKBvbaPl2C3C395OSITdtlJBXBzdHcmWovc5wU9aVBsNcdMAvA=="
+}
